# docs/swagger.yaml
info:
  contact: {}
paths:
  /articles:
    get:
      consumes:
      - application/json
      description: List all articles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: internal
      summary: List All Articles
      tags:
      - Articles
  /articles/:
    delete:
      consumes:
      - application/json
      description: Delete article by article id
      parameters:
      - description: the id of a specfic article
        in: query
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: invalid
      summary: Delete Article By Article Id
      tags:
      - Articles
    post:
      consumes:
      - application/json
      description: Create Article
      parameters:
      - description: Name
        in: query
        name: ID
        required: true
        type: string
      - description: Title
        in: query
        name: title
        type: integer
      - description: Content
        in: query
        name: content
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: invalid
      summary: Create Article
      tags:
      - Articles
  /articles/{articleId}:
    get:
      consumes:
      - application/json
      description: Get article by article id
      parameters:
      - description: the id of a specfic article
        in: path
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: invalid
      summary: Get Article By Article Id
      tags:
      - Articles
    put:
      consumes:
      - application/json
      description: Update article by article id
      parameters:
      - description: the id of a specfic article
        in: path
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: invalid
      summary: Update Article By Article Id
      tags:
      - Articles
swagger: "2.0"